+++ sources/src/index.ts
@@ -2,8 +2,6 @@ export * from "./keys";
 export * from "./types";
 
 export * from "./interceptors";
-export * from "./models";
-export * from "./repositories";
 export * from "./providers";
 export * from "./servers";
 export * from "./mixins";