+++ sources/src/interceptors/index.ts
@@ -1,4 +1,5 @@
+export * from "./utils";
+
 export * from "./validate.interceptor";
-export * from "./unique.interceptor";
+export * from "./limit.interceptor";
 export * from "./exist.interceptor";
-export * from "./filter.interceptor";