+++ sources/src/models/index.ts
@@ -1,9 +1,0 @@
-export * from "./user.model";
-export * from "./role.model";
-export * from "./permission.model";
-
-export * from "./user-role.model";
-export * from "./role-permission.model";
-
-export * from "./session.model";
-export * from "./code.model";