+++ sources/src/providers/index.ts
@@ -1,5 +1,1 @@
-export * from "./authentication/token.service";
 export * from "./authentication/token.strategy";
-
-export * from "./message.provider";
-export * from "./activate.provider";