+++ sources/src/repositories/permission.repository.ts
@@ -1,8 +1,0 @@
-import { PermissionRepositoryMixin } from "loopback-component-authorization";
-
-import { Permission, PermissionRelations } from "../models";
-
-export class DefaultPermissionRepository extends PermissionRepositoryMixin<
-    Permission,
-    PermissionRelations
->()() {}