+++ sources/src/servers/rest/server.ts
@@ -4,38 +4,29 @@ import {
     CoreBindings,
     Application,
 } from "@loopback/core";
-import { Ctor } from "loopback-component-history";
 
 /** Swagger binding imports */
 import { RestServer, RestComponent } from "@loopback/rest";
 import { RestExplorerComponent } from "@loopback/rest-explorer";
 
 /** Authentication binding imports */
-import { AuthenticationComponent } from "@loopback/authentication";
+import {
+    AuthenticationComponent,
+    registerAuthenticationStrategy,
+} from "@loopback/authentication";
 
-import { CRUDBindings, PrivateCRUDBindings } from "../../keys";
+/** Authorization binding imports */
+import {
+    AuthorizationOptions,
+    AuthorizationDecision,
+    AuthorizationBindings,
+    AuthorizationComponent,
+} from "@loopback/authorization";
+
+import { CRUDBindings } from "../../keys";
 import { CRUDRestServerConfig } from "../../types";
+import { CRUDTokenStrategy } from "../../providers";
 import { Sequence } from "../../servers";
-
-import {
-    User,
-    Role,
-    Permission,
-    UserRole,
-    RolePermission,
-    Session,
-    Code,
-} from "../../models";
-
-import {
-    GenerateUsersController,
-    GenerateUsersSelfController,
-    GenerateUsersSessionController,
-    GenerateUsersAccountController,
-    GenerateUsersPasswordController,
-    GenerateRolesController,
-    GeneratePermissionsController,
-} from "../../servers/rest/controllers";
 
 @lifeCycleObserver("servers.REST")
 export class CRUDRestServer extends RestServer {
@@ -43,21 +34,7 @@ export class CRUDRestServer extends RestServer {
         @inject(CoreBindings.APPLICATION_INSTANCE)
         app: Application,
         @inject(CRUDBindings.REST_SERVER_CONFIG)
-        config: CRUDRestServerConfig = {},
-        @inject(PrivateCRUDBindings.USER_MODEL)
-        userCtor: Ctor<User>,
-        @inject(PrivateCRUDBindings.ROLE_MODEL)
-        roleCtor: Ctor<Role>,
-        @inject(PrivateCRUDBindings.PERMISSION_MODEL)
-        permissionCtor: Ctor<Permission>,
-        @inject(PrivateCRUDBindings.USER_ROLE_MODEL)
-        userRoleCtor: Ctor<UserRole>,
-        @inject(PrivateCRUDBindings.ROLE_PERMISSION_MODEL)
-        rolePermissionCtor: Ctor<RolePermission>,
-        @inject(PrivateCRUDBindings.SESSION_MODEL)
-        sessionCtor: Ctor<Session>,
-        @inject(PrivateCRUDBindings.CODE_MODEL)
-        codeCtor: Ctor<Code>
+        config: CRUDRestServerConfig = {}
     ) {
         super(app, config);
 
@@ -71,6 +48,16 @@ export class CRUDRestServer extends RestServer {
 
         /** Bind authentication component */
         app.component(AuthenticationComponent);
+        registerAuthenticationStrategy(app, CRUDTokenStrategy);
+
+        /** Bind authorization component */
+        app.configure<AuthorizationOptions>(AuthorizationBindings.COMPONENT).to(
+            {
+                precedence: AuthorizationDecision.DENY,
+                defaultDecision: AuthorizationDecision.DENY,
+            }
+        );
+        app.component(AuthorizationComponent);
 
         /** Bind swagger component */
         app.component(RestComponent);
@@ -82,27 +69,6 @@ export class CRUDRestServer extends RestServer {
 
         /** Set up the custom sequence */
         this.sequence(Sequence);
-
-        /** Bind users controllers */
-        app.controller(GenerateUsersController<User>(userCtor));
-        app.controller(GenerateUsersSelfController<User>(userCtor));
-        app.controller(
-            GenerateUsersSessionController<Session, User>(sessionCtor, userCtor)
-        );
-        app.controller(
-            GenerateUsersAccountController<Code, User>(codeCtor, userCtor)
-        );
-        app.controller(
-            GenerateUsersPasswordController<Code, User>(codeCtor, userCtor)
-        );
-
-        /** Bind roles controllers */
-        app.controller(GenerateRolesController<Role>(roleCtor));
-
-        /** Bind permissions controllers */
-        app.controller(
-            GeneratePermissionsController<Permission>(permissionCtor)
-        );
     }
 
     async start() {