+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.5.0
+
+-   **Deps**: update packages
+-   **Refactor**: remove `history`, `authorization` packages
+-   **Feat**: add `exist`, `validate`,`limit` interceptors
+-   **Feat:** add `TargetsManyController`, `TargetsOneController`
+
 ## v0.3.0
 
 -   **Deps**: update `history`, `authorization` packages